move getting signups to correct place
